Add the filter package
